refactor: Move the code along.
Update ReadMe
Move PostCode api to its own file.
Move filereader to its own file.
Optomise imports.
Add basic file validation for required cols.
